packages/conversational-agent-external-history-pinecone/pyproject.toml
[tool.poetry]
name = "conversational-agent-external-history-pinecone"
version = "0.1.0"
description = ""
authors = ["Sebastian Montero <sebastian.montero@gmail.com>"]
readme = "README.md"
# packages = [{include = "conversational_rag"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
python-dotenv = "^1.0.1"
langchain-core = "^0.1.44"
langchain-community = "^0.0.33"
langchain-openai = "^0.1.3"
pydantic = "1.10.9"
asyncio = "^3.4.3"
cohere = "^5.3.2"
langchain-cohere = "^0.1.3"
pinecone-client = "^3.2.2"
langchain-pinecone = "^0.1.0"
langchainhub = "^0.1.15"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
